requirements.txt
# requirements.txt
streamlit==1.22.0
pandas==1.5.3
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==0.21.0
email-validator==1.3.1
sqlalchemy==2.0.15